refactor: properties 매핑 클래스 디렉토리 수정 (#42)
- DTO 내부에 properties 디렉토리로 옮김
